New forward struct impl
